bot/requirements.txt
python-telegram-bot==20.0
SQLAlchemy==2.0.0
psycopg2==2.9.5
